Version 1.5.1 批量删除功能的输入框添加滚动条，添加空文件夹删除检测，UI 优化

+++ src/main/java/github/kasuminova/fileutils/FileThread.java
@@ -85,19 +85,21 @@ public class FileThread {
          * @param file 文件夹
          */
         public static void delDir(File file) {
-            // 判断是否是一个目录, 不是的话跳过, 直接删除; 如果是一个目录, 先将其内容清空.
-            if (file.isDirectory()) {
-                // 获取子文件/目录
-                File[] subFiles = file.listFiles();
-                // 遍历该目录
-                for (File subFile : subFiles) {
-                    // 递归调用删除该文件: 如果这是一个空目录或文件, 一次递归就可删除. 如果这是一个非空目录, 多次
-                    // 递归清空其内容后再删除
-                    delDir(subFile);
+            if (file.exists()) {
+                // 判断是否是一个目录, 不是的话跳过, 直接删除; 如果是一个目录, 先将其内容清空.
+                if (file.isDirectory()) {
+                    // 获取子文件/目录
+                    File[] subFiles = file.listFiles();
+                    // 遍历该目录
+                    for (File subFile : subFiles) {
+                        // 递归调用删除该文件: 如果这是一个空目录或文件, 一次递归就可删除. 如果这是一个非空目录, 多次
+                        // 递归清空其内容后再删除
+                        delDir(subFile);
+                    }
                 }
+                // 删除空目录或文件
+                file.delete();
             }
-            // 删除空目录或文件
-            file.delete();
         }
 
         /**
